% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity_types}
\alias{activity_types}
\title{Codes for Activity Types}
\format{
\subsection{\code{activity_types}}{

A data frame with 125 rows and 2 columns:
\describe{
\item{code}{Thryve ID specifying a type of activity for ActivityTypeDetail1 and ActivityTypeDetail2}
\item{activity}{The type of activity, e.g. "Walk", "Run, "Bike", etc.}
}
}
}
\source{
\url{https://docs-old.thryve.health/biomarkers.php#activitytypedetail1}
}
\usage{
activity_types
}
\description{
A table containing the connection between the Thryve ID codes for activity types listed in the
\code{value} column for variables ActivityTypeDetail1 and ActivityTypeDetail2.
}
\keyword{datasets}
