% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_mpathwear.R
\name{clean_daily_data}
\alias{clean_daily_data}
\title{Clean mpathwear daily data}
\usage{
clean_daily_data(
  data,
  .col = "dailyData",
  connectionId = "connectionId",
  start = "startUTS",
  end = "stopUTS"
)
}
\arguments{
\item{data}{A data frame containing the wearable data, as returned by \code{\link[=read_mpathwear]{read_mpathwear()}}.}

\item{.col}{The column containing the data to be unpacked.}

\item{connectionId}{The column containing the participant ID.}

\item{start}{The column containing the start time of the data retrieval period.}

\item{end}{The column containing the end time of the data retrieval period.}
}
\value{
A tibble containing at least the following columns:
\tabular{ll}{
\code{connectionId} \tab The participant ID in m-Path, as specified by the \code{connectionId} argument.\cr
\code{day} \tab The day of the measurement as a Date object. \cr
\code{category} \tab The overall category of the measurement. \cr
\code{subcategory} \tab The more specific category of the measurement. \cr
\code{variable} \tab The variable name of the measurement. \cr
\code{value} \tab The value of the measurement. \cr
\code{timezoneOffset} \tab The timezone offset of the measurement compared to UTC. \cr
\code{generation} \tab Whether the measurement was  a calculation, automatic measurement, or manual
input. \cr
\code{trustworthiness} \tab Whether the measurement was trustworthy. \cr
\code{created_at} \tab The time the measurement was created or updated. \cr
\code{data_source} \tab The data source of the measurement. Also see \code{\link[=data_sources]{data_sources()}}. \cr
\code{day_complete} \tab A flag that signals whether the day was complete, i.e. a value of 1. A value
of 0 indicates that it is uncertain that the day was complete, as there was no data the next
day. \cr
\code{description} \tab The description of the measurement. \cr
\code{available_sources} \tab The available data sources (e.g. Garmin, Fitbit) for this type of
measurement. \cr
}
}
\description{
Clean mpathwear daily data
}
