% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_mpathwear.R
\name{clean_dynamic_data}
\alias{clean_dynamic_data}
\title{Clean mpathwear intraday (i.e. dynamic) data}
\usage{
clean_dynamic_data(data, .col = "dynamicData", connectionId = "connectionId")
}
\arguments{
\item{data}{A data frame containing the wearable data, as returned by \code{\link[=read_mpathwear]{read_mpathwear()}}.}

\item{.col}{The column containing the data to be unpacked.}

\item{connectionId}{The column containing the participant ID.}
}
\value{
A tibble containing at least the following columns:
\tabular{ll}{
\code{connectionId} \tab The participant ID in m-Path, as specified by the \code{connectionId} argument.\cr
\code{start_time} \tab The start time of the measurement. \cr
\code{end_time} \tab The end time of the measurement. \cr
\code{category} \tab The overall category of the measurement. \cr
\code{subcategory} \tab The more specific category of the measurement. \cr
\code{variable} \tab The variable name of the measurement. \cr
\code{value} \tab The value of the measurement. \cr
\code{timezoneOffset} \tab The timezone offset of the measurement compared to UTC. \cr
\code{generation} \tab Whether the measurement was  a calculation, automatic measurement, or manual
input. \cr
\code{chronologicalExactness} \tab The imprecision of the timestamp in minutes. This imprecision
occurs when some specific sources give the summary of activity without precise information about
timestamp of specific activity. \cr
\code{created_at} \tab The time the measurement was created or updated. \cr
\code{data_source} \tab The data source of the measurement. Also see \link{data_sources}. \cr
\code{description} \tab The description of the measurement. \cr
\code{available_sources} \tab The available data sources (e.g. Garmin, Fitbit) for this type of
measurement. \cr
}
}
\description{
Clean mpathwear intraday (i.e. dynamic) data
}
