% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Overview of Thryve measurement codes and descriptions}
\format{
\subsection{\code{codebook}}{

A data frame with 344 rows and 8 columns:
\describe{
\item{category}{The overall category of the measurement.}
\item{subcategory}{The more specific category of the measurement.}
\item{level}{The measurement of the variable, i.e. \code{Daily} or \code{Intraday}}
\item{variable}{The variable name of the measurement.}
\item{description}{The description of the measurement.}
\item{code}{The Thryve code for the measurement, as used in the \code{dailyDynamicValueType}
and \code{dynamicValueType} columns of the mpathwear data.}
\item{unit}{The unit of the measurement, as used in the \code{valueType} column of the mpathwear
data.}
\item{available_sources}{The available data sources (e.g. Garmin, Fitbit) for this type of
measurement.}
}
}
}
\source{
\url{https://docs-old.thryve.health/biomarkers.php}

\url{https://docs-old.thryve.health/analytics.php}
}
\usage{
codebook
}
\description{
A table containing the connection between the Thryve ID codes for measurements listed in the
\code{dailyValueType} and \code{dynamicValueType} columns of the mpathwear data, along with descriptions,
units, and available data sources.
}
\keyword{datasets}
