% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sources}
\alias{data_sources}
\title{Overview of supported Thryve data sources and data retrieval periods}
\format{
\subsection{\code{data_sources}}{

A data frame with 38 rows and 5 columns:
\describe{
\item{id}{ID number of the data source in Thryve.}
\item{data_source}{The name of the data source.}
\item{type_integration}{How Thryve connects to the data source, e.g. oAuth (via API).}
\item{data_retrieval_frequency}{How and how often data is pulled from the data source.}
\item{type_data_source}{The status of integration of the data sources in Thryve. One of
"active", "experimental", or "thryve".}
}
}
}
\source{
\url{https://docs-old.thryve.health/access.php#data-sources}
}
\usage{
data_sources
}
\description{
A table containing the connection between the Thryve ID codes for activity types listed in the
\code{value} column for variables ActivityTypeDetail1 and ActivityTypeDetail2.
}
\keyword{datasets}
