% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpathwear-package.R
\docType{package}
\name{mpathwear-package}
\alias{mpathwear}
\alias{mpathwear-package}
\title{mpathwear: Imports and visualises wearable data from 'm-Path'}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Koen Niemeijer \email{koen.niemeijer@kuleuven.be} (\href{https://orcid.org/0000-0002-0816-534X}{ORCID})

}
\keyword{internal}
