% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpathwear.R
\name{read_mpathwear}
\alias{read_mpathwear}
\title{Read wearable data from m-Path}
\usage{
read_mpathwear(path, recursive = FALSE)
}
\arguments{
\item{path}{A character string to the directory containing the data files or a single file path.}

\item{recursive}{A logical value indicating whether data files should be searched in
subdirectories.}
}
\value{
A tibble containing the following columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Explanation} \cr
   connectionId \tab The participant ID in m-Path. \cr
   legacyCode \tab The legacy invitation code format. Corresponds directly to \code{code}. \cr
   code \tab The invitation code used in the study, if any. \cr
   alias \tab The alias of the participant (can be changed by the participant at any point). \cr
   initials \tab The initials of the participant based on the first seen alias. \cr
   accountCode \tab The account code of the researcher. \cr
   startUTS \tab The start of the requested data period in Unix timestamp format. \cr
   stopUTS \tab The end of the requested data period in Unix timestamp format. \cr
   lastCreatedAtUnix \tab The last Unix timestamp at which the wearable data was updated. \cr
   dynamicData \tab The intraday (i.e. momentary) wearable data. \cr
   dailyData \tab The interday (i.e. daily) wearable data. \cr
}


See \href{https://m-path.io/manual/knowledge-base/export-data/}{the m-Path manual page on data exporting}
for more information on the columns \code{connectionId} through \code{accountCode}, as these are
part of the default m-Path data output.
}
\description{
This function reads one more multiple CSV files containing wearable data exported from m-Path.
}
\examples{
# Your path to the data, or in this case the package example data.
# Note that this can also be a folder containing several files.
path <- system.file("extdata", "example.csv", package = "mpathwear")

read_mpathwear(path)
}
\seealso{
\code{\link[=clean_dynamic_data]{clean_dynamic_data()}} and \code{\link[=clean_daily_data]{clean_daily_data()}} for unpacking the dynamic and daily data.
}
