% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_chart.R
\name{sleep_rem_duration}
\alias{sleep_rem_duration}
\title{Calculate total REM sleep duration}
\usage{
sleep_rem_duration(
  .data,
  start = "start_time",
  end = "end_time",
  variable = "variable",
  tz_offset = "tz_offset"
)
}
\arguments{
\item{.data}{A data frame containing the wearable data, typically from \code{\link[=clean_dynamic_data]{clean_dynamic_data()}}.}

\item{start}{The name of the column containing start timestamps. Defaults to \code{"start_time"}.}

\item{end}{The name of the column containing end timestamps. Defaults to \code{"end_time"}.}

\item{variable}{The name of the column containing variable names. Defaults to \code{"variable"}.}

\item{tz_offset}{The name of the column containing timezone offsets. Defaults to \code{"tz_offset"}.}
}
\value{
A data frame with columns for \code{day} and \code{SleepREMDuration} (in seconds).
}
\description{
Calculates the total time spent in REM sleep for each night from intraday wearable data.
}
\examples{
#' # Calculate the total REM duration during sleep from
# intraday (dynamic) data.
sleep_rem_duration(dynamic_data)

# We can compare this to the REM duration from the
# daily data.
# Note that in the daily data, the REM duration is shown
# in minutes instead of seconds.
daily_data[daily_data$variable == "SleepREMDuration", c("day", "value")]
}
\seealso{
\code{\link[=sleep_deep_duration]{sleep_deep_duration()}}, \code{\link[=sleep_duration]{sleep_duration()}}, \code{\link[=sleep_chart]{sleep_chart()}}
}
