% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_overlapping_time}
\alias{summarise_overlapping_time}
\title{Summarise Overlapping and Contiguous Time Intervals}
\usage{
summarise_overlapping_time(.data, start, end, tolerance = 0)
}
\arguments{
\item{.data}{A data frame containing at least the \code{start} and \code{end} time columns.}

\item{start}{The name of the column indicating start times. Defaults to \code{"start_time"}.}

\item{end}{The name of the column indicating end times. Defaults to \code{"end_time"}.}

\item{tolerance}{A numeric value indicating the maximum difference between the \code{end} of one interval
and the \code{start} of the next interval to consider them contiguous. Defaults to \code{0}.}
}
\value{
A data frame with collapsed contiguous time intervals, where each row represents a continuous span.
}
\description{
Summarise rows in a dataset where the \code{start} of the next interval is (nearly) equal to the
\code{end} of the current interval, indicating they are contiguous. This is useful for simplifying
sequences of adjacent time intervals into broader continuous spans. It also summarises over time
intervals if they overlap.
}
\examples{
df <- data.frame(
  start_time = c(1, 2, 4, 5),
  end_time = c(2, 3, 5, 6)
)
summarise_overlapping_time(
  df,
  start = start_time,
  end = end_time
)

}
