% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpathwear-package.R
\docType{package}
\name{mpathwear-package}
\alias{mpathwear}
\alias{mpathwear-package}
\title{mpathwear: Imports and visualises wearable data from 'm-Path'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/koenniem/mpathwear}
  \item \url{https://koenniem.github.io/mpathwear/}
  \item Report bugs at \url{https://github.com/koenniem/mpathwear/issues}
}

}
\author{
\strong{Maintainer}: Koen Niemeijer \email{koen.niemeijer@kuleuven.be} (\href{https://orcid.org/0000-0002-0816-534X}{ORCID})

}
\keyword{internal}
